requirements.txt
pygame~=2.0.1
neat-python